[PrestoOnSpark] Extract interface from GroupedTopNBuilder
