+++ presto-main/src/test/java/com/facebook/presto/operator/BenchmarkInMemoryGroupedTopNBuilder.java
@@ -58,7 +58,7 @@ import static org.testng.Assert.assertFalse;
 @Fork(4)
 @Warmup(iterations = 10, time = 500, timeUnit = TimeUnit.MILLISECONDS)
 @Measurement(iterations = 10, time = 500, timeUnit = TimeUnit.MILLISECONDS)
-public class BenchmarkGroupedTopNBuilder
+public class BenchmarkInMemoryGroupedTopNBuilder
 {
     private static final int HASH_GROUP = 0;
     private static final int EXTENDED_PRICE = 1;
@@ -90,7 +90,7 @@ public class BenchmarkGroupedTopNBuilder
         private int groupCount = 10;
 
         private List<Page> page;
-        private GroupedTopNBuilder topNBuilder;
+        private InMemoryGroupedTopNBuilder topNBuilder;
 
         @Setup
         public void setup()
@@ -99,7 +99,7 @@ public class BenchmarkGroupedTopNBuilder
             GroupByHash groupByHash;
             if (groupCount > 1) {
                 groupByHash = new BigintGroupByHash(HASH_GROUP, true, groupCount, UpdateMemory.NOOP);
-                topNBuilder = new GroupedTopNBuilder(
+                topNBuilder = new InMemoryGroupedTopNBuilder(
                         null,
                         types,
                         ImmutableList.of(types.get(HASH_GROUP)),
@@ -114,7 +114,7 @@ public class BenchmarkGroupedTopNBuilder
                         UpdateMemory.NOOP);
             }
             else {
-                topNBuilder = new GroupedTopNBuilder(
+                topNBuilder = new InMemoryGroupedTopNBuilder(
                         null,
                         types,
                         Collections.emptyList(),
@@ -130,7 +130,7 @@ public class BenchmarkGroupedTopNBuilder
             }
         }
 
-        public GroupedTopNBuilder getTopNBuilder()
+        public InMemoryGroupedTopNBuilder getTopNBuilder()
         {
             return topNBuilder;
         }
@@ -144,7 +144,7 @@ public class BenchmarkGroupedTopNBuilder
     @Benchmark
     public void topN(BenchmarkData data, Blackhole blackhole)
     {
-        GroupedTopNBuilder topNBuilder = data.getTopNBuilder();
+        InMemoryGroupedTopNBuilder topNBuilder = data.getTopNBuilder();
         for (Page page : data.getPages()) {
             Work<?> work = topNBuilder.processPage(page);
             boolean finished;
@@ -161,7 +161,7 @@ public class BenchmarkGroupedTopNBuilder
 
     public List<Page> topNToList(BenchmarkData data)
     {
-        GroupedTopNBuilder topNBuilder = data.getTopNBuilder();
+        InMemoryGroupedTopNBuilder topNBuilder = data.getTopNBuilder();
         for (Page page : data.getPages()) {
             Work<?> work = topNBuilder.processPage(page);
             boolean finished;
@@ -185,7 +185,7 @@ public class BenchmarkGroupedTopNBuilder
     {
         Options options = new OptionsBuilder()
                 .parent(new CommandLineOptions(args))
-                .include(".*" + BenchmarkGroupedTopNBuilder.class.getSimpleName() + ".*")
+                .include(".*" + BenchmarkInMemoryGroupedTopNBuilder.class.getSimpleName() + ".*")
                 .build();
 
         new Runner(options).run();
